feat(signup): extract shared component
