+++ pages/signup.vue
@@ -17,164 +17,10 @@
             </h4>
           </div>
 
-          <FormKit
-            type="form"
+          <ProfileForm
             @submit="handleSubmit"
-            :submit-label="$t('common.submit')"
-            :submit-attrs="{
-              wrapperClass: 'signup-form_elm-submit',
-            }"
-            :config="{
-              classes: {
-                outer: 'signup-form_outer',
-                wrapper: 'c-form_formulate-wrapper',
-                fieldset: 'c-form_formulate-fieldset',
-              },
-            }"
-          >
-            <div class="signup-form_container">
-              <FormKit
-                v-model="sitekey"
-                name="sitekey"
-                :label="$t('login.site_key')"
-                type="text"
-                :classes="{ outer: 'signup-form_elm-sitekey' }"
-                validation="required|length:0,100"
-              />
-              <FormKitSchema
-                :schema="[
-                  {
-                    name: 'name1',
-                    label: $t('label.first_name'),
-                    $formkit: 'text',
-                    validation: 'required|length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-name1',
-                    },
-                  },
-                  {
-                    name: 'name2',
-                    label: $t('label.last_name'),
-                    $formkit: 'text',
-                    validation: 'required|length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-name2',
-                    },
-                  },
-                  {
-                    name: 'hire_date',
-                    label: $t('label.hire_date'),
-                    $formkit: 'date',
-                    validation: 'required',
-                    classes: {
-                      outer: 'signup-form_elm-hire_date',
-                    },
-                  },
-                  {
-                    name: 'department',
-                    label: $t('label.department'),
-                    $formkit: 'text',
-                    validation: 'length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-department',
-                    },
-                  },
-                  {
-                    name: 'position',
-                    label: $t('label.position'),
-                    $formkit: 'text',
-                    validation: 'length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-position',
-                    },
-                  },
-                  {
-                    name: 'tel',
-                    label: $t('label.phone'),
-                    $formkit: 'text',
-                    validation: 'required|length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-tel',
-                    },
-                  },
-                  {
-                    name: 'email',
-                    label: $t('label.email'),
-                    $formkit: 'email',
-                    validation: 'required|email|length:0,100',
-                    classes: {
-                      outer: 'signup-form_elm-email',
-                    },
-                  },
-                  {
-                    name: 'login_pwd',
-                    label: $t('label.password'),
-                    $formkit: 'password',
-                    validation: 'required|length:8,100',
-                    placeholder: 'Password',
-                    classes: {
-                      outer: 'signup-form_elm-login_pwd',
-                    },
-                  },
-                  {
-                    name: 'pull_down',
-                    label: $t('label.office'),
-                    $formkit: 'select',
-                    placeholder: 'Select office',
-                    options: [
-                      { label: '', value: '' },
-                      { label: 'Tokyo', value: '1' },
-                      { label: 'Osaka', value: '2' },
-                      { label: 'Malaysia', value: '3' },
-                    ],
-                    classes: {
-                      outer: 'signup-form_elm-office',
-                    },
-                  },
-                  {
-                    name: 'multiple_check',
-                    label: $t('label.hobby'),
-                    $formkit: 'checkbox',
-                    options: [
-                      { label: 'Reading', value: '1' },
-                      { label: 'Watching TV', value: '2' },
-                      { label: 'Family Time', value: '3' },
-                      { label: 'Going to Movies', value: '4' },
-                      { label: 'Fishing', value: '5' },
-                      { label: 'Computer', value: '6' },
-                      { label: 'Gardening', value: '7' },
-                      { label: 'Renting Movies', value: '8' },
-                      { label: 'Walking', value: '9' },
-                      { label: 'Excercise', value: '10' },
-                    ],
-                    classes: {
-                      outer: 'signup-form_elm-hobby',
-                    },
-                  },
-                  {
-                    name: 'notes',
-                    label: $t('label.notes'),
-                    $formkit: 'textarea',
-                    validation: 'length:0,1000',
-                    classes: {
-                      outer: 'signup-form_elm-notes',
-                    },
-                  },
-                ]"
-              ></FormKitSchema>
-            </div>
-            <FormKit
-              type="checkbox"
-              :label="$t('common.agree')"
-              name="term"
-              v-model="agreementChecked"
-              validation="accepted"
-              :classes="{
-                outer: 'signup-form_elm-terms',
-                wrapper: 'signup-form_elm-terms_wrapper',
-              }"
-            />
-          </FormKit>
+          />
+
         </v-container>
       </div>
     </div>
@@ -186,9 +32,8 @@ const snackbar = useSnackbar();
 const { register } = useAuth();
 const loading = ref(false);
 const sitekey = ref(apiDomain.sitekey);
-const agreementChecked = ref(false);
 
-const handleSubmit = async (form) => {
+const handleSubmit = async (formValue) => {
   loading.value = true;
 
   apiDomain.sitekey = sitekey.value;
@@ -199,7 +44,7 @@ const handleSubmit = async (form) => {
 
   try {
     await register({
-      ...form,
+      ...formValue,
       // remove unnecessary fields
       sitekey: undefined,
       term: undefined,
@@ -288,4 +133,5 @@ const handleSubmit = async (form) => {
     }
   }
 }
-</style>
+</style>import type { ProfileForm } from '#build/components';
+
